+++ src/pydidas_plugins/input_plugins/fio_mca_line_scan_series_loader.py
@@ -41,6 +41,9 @@ from pydidas.core import (
 from pydidas.core.utils import copy_docstring
 from pydidas.core.utils import fio_utils as fio
 from pydidas.plugins import InputPlugin, InputPlugin1d
+from pydidas.widgets.plugin_config_widgets.plugin_config_widget_with_custom_xscale import (
+    PluginConfigWidgetWithCustomXscale,
+)
 
 
 SCAN = ScanContext()
@@ -48,7 +51,7 @@ SCAN = ScanContext()
 
 class FioMcaLineScanSeriesLoader(InputPlugin1d):
     """
-    Load data frames from a series of Fio files with MCA data.
+    Load 1d data from a series of Fio files with MCA data.
 
     This plugin is designed to allow loading .fio files written by Sardana which
     include a single row of data with the MCA spectrum.
@@ -114,9 +117,11 @@ class FioMcaLineScanSeriesLoader(InputPlugin1d):
         "files_per_directory",
         "_counted_files_per_directory",
         "fio_suffix",
-        "use_absolute_energy",
-        "energy_offset",
-        "energy_delta",
+        "use_custom_xscale",
+        "x0_offset",
+        "x_delta",
+        "x_label",
+        "x_unit",
     )
 
     def __init__(self, *args: tuple, **kwargs: dict):
@@ -175,14 +180,6 @@ class FioMcaLineScanSeriesLoader(InputPlugin1d):
                 self.get_param_value("files_per_directory"),
             )
 
-    def _determine_header_size(self):
-        """
-        Determine the size of the header in lines.
-        """
-        _n_header, _n_data = fio.determine_header_and_data_lines(self.get_filename(0))
-        self._config["header_lines"] = _n_header
-        self._config["data_lines"] = _n_data
-
     def get_frame(self, index: int, **kwargs: dict) -> tuple[Dataset, dict]:
         """
         Get the frame for the given index.
@@ -201,7 +198,7 @@ class FioMcaLineScanSeriesLoader(InputPlugin1d):
         kwargs : dict
             The updated kwargs.
         """
-        _dataset = fio.load_fio_energy_spectrum(self.get_filename(index), self._config)
+        _dataset = fio.load_fio_spectrum(self.get_filename(index), self._config)
         return _dataset, kwargs
 
     @copy_docstring(InputPlugin)
@@ -219,3 +216,7 @@ class FioMcaLineScanSeriesLoader(InputPlugin1d):
         )
         _file_index = index % _n_per_dir + 1
         return self.filename_string.format(index0=_path_index, index1=_file_index)
+
+    def get_parameter_config_widget(self):
+        """Get the parameter config widget for the plugin."""
+        return PluginConfigWidgetWithCustomXscale